+++ src/main/java/me/xjqsh/lrtactical/entity/EffectCloudGrenadeEntity.java
@@ -81,7 +81,7 @@ public class EffectCloudGrenadeEntity extends ThrowableItemEntity {
 
     @Override
     protected Item getDefaultItem() {
-        return me.xjqsh.lrtactical.init.ModItems.THROWABLE;
+        return me.xjqsh.lrtactical.init.ModItems.THROWABLE.get();
     }
 
     @Nullable